BUG/MINOR: pool/stats: Use ullong to report total pool usage in bytes in stats
The same change was already performed for the cli. The stats applet and the
prometheus exporter are also concerned. Both use the stats API and rely on
pool functions to get total pool usage in bytes. pool_total_allocated() and
pool_total_used() must return 64 bits unsigned integer to avoid any wrapping
around 4G.

This may be backported to all versions.

(cherry picked from commit c960a3b60f5d05b82cdac2a33ab22ca465787e60)
(cherry picked from commit b174d82dff11d7fb67e9a7f53c20a658f23dd9e7)
(cherry picked from commit a6ff9f5361894e6e75fe26ffe5fce4b98fb4bee2)
(cherry picked from commit 3974420fb1bbfcc28bd9fcbd70903d72732adf5f)
(cherry picked from commit 7b9b3d289b754118af8d08f72256b5b20aa2c128)
[cf: No change in src/stats.c and ullong does not exist]
(cherry picked from commit bf9b7e38281465a9edc827e1bf0b8d9ce815f0ea)

+++ include/common/memory.h
@@ -131,8 +131,8 @@ void create_pool_callback(struct pool_head **ptr, char *name, unsigned int size)
 void dump_pools_to_trash();
 void dump_pools(void);
 int pool_total_failures();
-unsigned long pool_total_allocated();
-unsigned long pool_total_used();
+unsigned long long pool_total_allocated();
+unsigned long long pool_total_used();
 
 /*
  * This function frees whatever can be freed in pool <pool>.
